server/src/main/java/server/sockets/AssignPrincipalHandshakeHandler.java
package server.sockets;

import org.springframework.http.server.ServerHttpRequest;
import org.springframework.web.socket.WebSocketHandler;
import org.springframework.web.socket.server.support.DefaultHandshakeHandler;

import java.security.Principal;
import java.util.Map;
import java.util.UUID;

/**
 * https://stackoverflow.com/questions/56497055/principal-is-null-for-every-spring-websocket-event
 * Assign a random username as principal for each websocket client. This is
 * needed to be able to communicate with a specific client.
 */

public class AssignPrincipalHandshakeHandler extends DefaultHandshakeHandler {
    private static final String ATTR_PRINCIPAL = "__principal__";

    @Override
    protected Principal determineUser(ServerHttpRequest request, WebSocketHandler wsHandler, Map<String, Object> attributes) {
        final String name;
        if (!attributes.containsKey(ATTR_PRINCIPAL)) {
            name = generateRandomUsername();
            attributes.put(ATTR_PRINCIPAL, name);
        } else {
            name = (String) attributes.get(ATTR_PRINCIPAL);
        }
        return new Principal() {
            @Override
            public String getName() {
                return name;
            }
        };
    }

    /**
     * Generate a random ID for the new user which connects
     * @return the randomly generated ID
     */
    public String generateRandomUsername() {
        return UUID.randomUUID().toString();
    }
}